move BooksForm, BooksLint and Books classes to containers folder
